+++ HtmlRun.Common.Runtime/Interfaces/IRuntimeContext.cs
@@ -5,11 +5,15 @@ namespace HtmlRun.Runtime.Interfaces;
 
 public interface IRuntimeContext : IBaseContext
 {
+  IHeap Heap { get; }
+
   List<string> Usings { get; }
 
   void AddUsing(string namesp);
-  
+
   void AddVariable(ContextValue value);
 
   ICurrentInstructionContext Fork(IHtmlRuntimeForContext runtimeForContext, string callName, IEnumerable<ParsedArgument> args);
+
+  void Release();
 }