+++ HtmlRun.Runtime/RuntimeContext/BaseContext.cs
@@ -4,4 +4,10 @@ namespace HtmlRun.Runtime.RuntimeContext;
 
 public abstract class BaseContext
 {
+  public string UUID { get; private set; }
+
+  public BaseContext()
+  {
+    this.UUID = Guid.NewGuid().ToString();
+  }
 }